[feat] #5 오늘의 작가 api 추가 및 test 코드 분리
